requirements.txt
google-genai
pytest